move outcomes processor
